% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baixar_deflator_anual.R
\name{baixar_deflator_anual}
\alias{baixar_deflator_anual}
\title{Download and create the deflator file for the Annual PNADC as both rds and csv}
\usage{
baixar_deflator_anual(destination_path)
}
\arguments{
\item{destination_path}{The folder in which the files are to be stored (if folder does not exist, it will be created)}

\item{ano}{Year of of the Annual PNADC}
}
\value{
The files in the designated path (rds and csv)
}
\description{
Download and create the deflator file for the Annual PNADC as both rds and csv
}
\examples{
baixar_deflator_anual(destination_path = "./deflator");
}
