% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baixar_leitor_anual.R
\name{baixar_leitor_anual}
\alias{baixar_leitor_anual}
\title{Download and create the reader file for the Annual PNADC as both rds and csv}
\usage{
baixar_leitor_anual(ano, destination_path)
}
\arguments{
\item{ano}{Year of of the Annual PNADC}

\item{destination_path}{The folder in which the files are to be stored (if folder does not exist, it will be created)}
}
\value{
The files in the designated path (rds and csv)
}
\description{
Download and create the reader file for the Annual PNADC as both rds and csv
}
\examples{
baixar_leitor_anual(ano = 2019, destination_path = "./leitores");
}
