% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baixar_pnadc_anual.R
\name{baixar_pnadc_anual}
\alias{baixar_pnadc_anual}
\title{Download and unzip Annual PNADC microdata}
\usage{
baixar_pnadc_anual(ano, destination_path)
}
\arguments{
\item{ano}{Year of of the Annual PNADC}

\item{destination_path}{The folder in which the files are to be stored (if folder does not exist, it will be created)}
}
\value{
The microdata files (txt) in the designated path
}
\description{
Download and create the reader file and save as both rds and csv
}
\examples{
baixar_pnadc_anual(ano = 2019, destination_path = "./microdata");
}
