% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular_desocupados_anual.R
\name{calcular_desocupados_anual}
\alias{calcular_desocupados_anual}
\title{Calcular desocupados}
\usage{
calcular_desocupados_anual(df_pnadc)
}
\arguments{
\item{df_pnadc}{O dataframe com dados da PNADC anual (ver ler_pnadc_anual)}
}
\value{
Um dataframe com 3 colunas: ano, total desocupados e taxa de desocupação (% da força de trabalho)
}
\description{
Calcular desocupados (número absoluto e taxa de desocupação)
}
\examples{
calcular_desocupados_anual(pnadc_2019);
}
\seealso{
ler_pnadc_anual
}
