% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular_informalidade.R
\name{calcular_informalidade}
\alias{calcular_informalidade}
\title{Calcular informalidade}
\usage{
calcular_informalidade(df_pnadc)
}
\arguments{
\item{df_pnadc}{O dataframe com dados da PNADC anual (ver ler_pnadc_anual)}
}
\value{
Um dataframe com 3 colunas: ano, total de informais, taxa de informalidade (% da população ocupada)
}
\description{
Calcular a informalidade: soma de trabalhadores do setor privado e domésticos sem carteira, empregados e por conta própria sem CNPJ e trabalhadores familiares auxiliares
}
\examples{
calcular_informalidade(pnadc_2018);
}
\references{
https://biblioteca.ibge.gov.br/visualizacao/livros/liv101743_informativo.pdf nota 7

Roubaud et al, 2020. Disponível em https://www.ie.ufrj.br/images/IE/TDS/2020/TD_IE_031_2020_ROUBAUD_et%20al.pdf
}
\seealso{
ler_pnadc_anual, montar_tabela_informalidade
}
