% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular_nem_nem_anual.R
\name{calcular_nem_nem_anual}
\alias{calcular_nem_nem_anual}
\title{Calcular população jovem nem-nem}
\usage{
calcular_nem_nem_anual(df_pnad, idade_min = 15, idade_max = 29)
}
\arguments{
\item{idade_min}{O patamar inferior de idade que define "jovem" (default = 15; mínimo = 14). A idade escolhida faz parte da população jovem}

\item{idade_max}{O patamar inferior de idade que define "jovem" (default = 15). A idade escolhida faz parte da população jovem}

\item{df_pnadc}{O dataframe com dados da PNADC anual (ver ler_pnadc_anual)}
}
\value{
Um dataframe com 3 colunas: ano, total nem-nem e % de nem-nem dentro da população jovem
}
\description{
Calcular população jovem nem-nem: pessoas com idade entre "idade_min" e "idade_max" que não estudam nem trabalham
}
\examples{
calcular_nem_nem_anual(pnadc_2013, idade_min = 15, idade_max = 29);
calcular_nem_nem_anual(pnadc_2014, idade_min = 14, idade_max = 18);
}
\references{
Neri, 2021 - https://www.cps.fgv.br/cps/NemNem/
}
\seealso{
ler_pnadc_anual
}
