% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular_pobreza_anual.R
\name{calcular_pobreza_anual}
\alias{calcular_pobreza_anual}
\title{Calculate poverty}
\usage{
calcular_pobreza_anual(df_pnadc, linha_pob)
}
\arguments{
\item{df_pnadc}{The dataframe with the Annual PNADC data (see ler_pnadc_anual)}

\item{linha_pob}{The monetary poverty threshold (household per capita monthly income)}
}
\value{
A dataframe with 6 cols: year, povery threshold, number of poor, P0 (proportion of poor - %), P1 (income-gap measure) and P2 (square income-gap)
}
\description{
From a poverty threshold, calculate the number of poor and the 3 indexes from the FGT family
}
\examples{
calcular_pobreza_anual(pnadc_2019, 400);
}
\references{
Foster, J.;  Greer, J.; Thorbecke, E. (1984) A class of decomposable poverty measures. Econometrica, v. 52, n. 3, pp. 761-766

See also https://guilhermejacob.github.io/context/2-poverty.html#poverty
}
