% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular_rdpc_media_anual.R
\name{calcular_rdpc_media_anual}
\alias{calcular_rdpc_media_anual}
\title{Calculate average household per capita monthly income}
\usage{
calcular_rdpc_media_anual(df_pnadc)
}
\arguments{
\item{df_pnadc}{The dataframe with the Annual PNADC data (see ler_pnadc_anual)}
}
\value{
A dataframe with 2 cols: year and the average household per capita monthly income
}
\description{
Calculate average household per capita monthly income
}
\examples{
calcular_rdpc_media_anual(pnadc_2019);
}
