% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular_rdpc_percentil_anual.R
\name{calcular_rdpc_percentil_anual}
\alias{calcular_rdpc_percentil_anual}
\title{Calculate percentile household per capita monthly income}
\usage{
calcular_rdpc_percentil_anual(df_pnadc, percentil = 0.5)
}
\arguments{
\item{df_pnadc}{The dataframe with the Annual PNADC data (see ler_pnadc_anual)}

\item{percentil}{Numeric. The percentile (default = 0.5 : median). Can be more than one (see examples)}
}
\value{
A dataframe with 3 cols: year, percentile and the respective household per capita monthly income
}
\description{
Calculate percentile household per capita monthly income
}
\examples{
calcular_rdpc_percentil_anual(pnadc_2019);
calcular_rdpc_percentil_anual(pnadc_2015, percentil = c(0.1, 0.4, 0.9, 0.95))
}
