% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular_renda_trabalho_media_anual.R
\name{calcular_renda_trabalho_media_anual}
\alias{calcular_renda_trabalho_media_anual}
\title{Calcular a renda média mensal do trabalho}
\usage{
calcular_renda_trabalho_media_anual(df_pnadc, tipo)
}
\arguments{
\item{df_pnadc}{O dataframe com dados da PNADC anual (ver ler_pnadc_anual)}

\item{tipo}{"habitual" ou "efetivo". O primeiro é a renda que costuma receber, o segundo é a renda efetivamente recebida}
}
\value{
Um dataframe com 3 colunas: ano, renda média mensal do trabalho real e nominal
}
\description{
Calcular a renda média mensal do trabalho
}
\examples{
calcular_renda_trabalho_media_anual(pnadc_2019, tipo = "habitual");
calcular_renda_trabalho_media_anual(pnadc_2013, tipo = "efetivo");
}
\seealso{
calcular_rdpc_media_anual,  calcular_renda_trabalho_percentil_anual
}
