% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular_renda_trabalho_percentil_anual.R
\name{calcular_renda_trabalho_percentil_anual}
\alias{calcular_renda_trabalho_percentil_anual}
\title{Calcular o valor da renda do trabalho mensal associado a determinado percentil}
\usage{
calcular_renda_trabalho_percentil_anual(df_pnadc, percentil = 0.5, tipo)
}
\arguments{
\item{df_pnadc}{O dataframe com dados da PNADC anual (ver ler_pnadc_anual)}

\item{percentil}{O percentil (default = 0.5 : isto é, o padrão é retornar a renda mediana). Pode ser mais de um valor na mesma chamada (ver exemplos)}

\item{tipo}{"habitual" ou "efetivo". O primeiro é a renda que costuma receber, o segundo é a renda efetivamente recebida}
}
\value{
Um dataframe com 3 colunas: ano, percentil e a respectiva renda do trabalho mensal máxima
}
\description{
Calcular o valor da renda do trabalho mensal associado a determinado percentil (é o ponto de corte, não o valor médio)
}
\examples{
calcular_renda_trabalho_percentil_anual(pnadc_2019);
calcular_renda_trabalho_percentil_anual(pnadc_2015, percentil = c(0.1, 0.4, 0.9, 0.95))
}
