% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{deflator}
\alias{deflator}
\title{Deflatores da PNADC anual}
\format{
Um dataframe com 972 linhas e 7 variáveis
\describe{
  \item{Ano}{ano}
  \item{ID_deflator}{Identificador, combinando Ano (4 dígitos), Trimestre (1 dígito) e UF (2 dígitos). É a chave para juntar aos dataframes com os microdados da PNADC}
  \item{CO1}{deflator a preços médios do ano, valores habituais}
  \item{CO1e}{deflator a preços médios do ano, valores efetivos}
  \item{CO2}{deflator a preços médios do último ano, valores habituais}
  \item{CO2e}{deflator a preços médios do último ano, valores efetivos}
  \item{CO3}{deflator para cálculos relacionados à linha de pobreza (ODM/ODS)}
}
}
\usage{
deflator
}
\description{
Deflatores da PNADC anual
}
\seealso{
https://ftp.ibge.gov.br/Trabalho_e_Rendimento/Pesquisa_Nacional_por_Amostra_de_Domicilios_continua/Anual/Microdados/Visita/Documentacao_Geral/PNADcIBGE_Deflator_Anual_Visita.pdf
}
\keyword{datasets}
