% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{leitores_2013}
\alias{leitores_2013}
\title{Annual PNADC reader 2013}
\format{
A data frame with 247 rows and 4 variables:
\describe{
  \item{inicio}{numeric, beggining position of the variable in the microdata txt file}
  \item{fim}{numeric, ending position of the variable in the microdata txt file}
  \item{tamanho}{numeric, size of the variable in the microdata txt file}
  \item{variavel}{character, name of the variable in the microdata txt file}
}
}
\usage{
leitores_2013
}
\description{
Annual PNADC reader 2013
}
\keyword{datasets}
