% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{leitores_2015}
\alias{leitores_2015}
\title{Leitores da PNADC anual - 2015}
\format{
Um dataframe com 287 linhas e 4 variáveis
\describe{
  \item{inicio}{numeric, posição inicial da variável no arquivo txt de microdados}
  \item{fim}{numeric, posição final da variável no arquivo txt de microdados}
  \item{tamanho}{numeric, tamanho da variável no arquivo txt de microdados}
  \item{variavel}{character, nome da variável no arquivo txt de microdados}
}
}
\usage{
leitores_2015
}
\description{
Leitores da PNADC anual - 2015
}
\keyword{datasets}
