% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_pnadc_anual.R
\name{ler_pnadc_anual}
\alias{ler_pnadc_anual}
\title{Carregar os microdados da PNADC anual}
\usage{
ler_pnadc_anual(ano, caminho_pnadc)
}
\arguments{
\item{ano}{ano}

\item{caminho_pnadc}{A pasta na qual os arquivos de microdados estão guardados (em geral a mesma utilizada na função baixar_pnadc_anual)}
}
\value{
Um dataframe com todos os microdados
}
\description{
Carregar (para o R) os microdados da PNADC anual
}
\examples{
ler_pnadc_anual(2018, "./pasta_microdados");
}
