% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montar_tabela_rdpc_anual.R
\name{montar_tabela_rdpc_anual}
\alias{montar_tabela_rdpc_anual}
\title{Adicionar renda domiciliar per capita mensal real como uma coluna}
\usage{
montar_tabela_rdpc_anual(df_pnadc)
}
\arguments{
\item{df_pnadc}{O dataframe com dados da PNADC anual (ver ler_pnadc_anual)}
}
\value{
Um dataframe sem as linhas nas quais VD2002 igual a 15, 16 ou 17; e com 1 coluna adicional: "rdpc" com a renda domiciliar per capital mensal real
}
\description{
Adicionar renda domiciliar per capita mensal real como uma coluna
}
\examples{
montar_tabela_rdpc_anual(pnadc_2019);
}
\seealso{
calcular_rdpc_media_anual, calcular_rdpc_percentil_anual
}
