% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnadcpackage-package.R
\docType{package}
\name{pnadcpackage-package}
\alias{pnadcpackage}
\alias{pnadcpackage-package}
\title{pnadcpackage: PNADC (IBGE - Brazil) Data and Analysis}
\description{
This package aims to provide summary statistics and family-specific data concerning expenditure and income, available at the Pesquisa Nacional por Amostra Domiciliar Continua (PNADC - IBGE).
}
\keyword{internal}
