% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baixar_deflator_anual.R
\name{baixar_deflator_anual}
\alias{baixar_deflator_anual}
\title{Baixa e cria o arquivo de deflator para a PNADC anual, em rds e em csv}
\usage{
baixar_deflator_anual(caminho_pasta)
}
\arguments{
\item{caminho_pasta}{A pasta (o caminho para a pasta) na qual os arquivos serão guardados (se não existir, a pasta será criada)}
}
\value{
Os arquivos rds e csv na pasta designada
}
\description{
Baixa e cria o arquivo de deflator para a PNADC anual, em rds e em csv (é preciso acesso a internet)
}
\examples{
baixar_deflator_anual(caminho_pasta = "./deflator");
}
