% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baixar_leitor_anual.R
\name{baixar_leitor_anual}
\alias{baixar_leitor_anual}
\title{Baixa e cria o arquivo de leitor para a PNADC anual, em rds e csv}
\usage{
baixar_leitor_anual(ano, caminho_pasta)
}
\arguments{
\item{ano}{Ano da PNADC Anual}

\item{caminho_pasta}{A pasta (o caminho para a pasta) na qual os arquivos serão guardados (se não existir, a pasta será criada)}
}
\value{
Os arquivos rds e csv na pasta designada
}
\description{
Baixa e cria o arquivo de leitor para a PNADC anual, em rds e csv (é preciso acesso a internet)
}
\examples{
baixar_leitor_anual(ano = 2019, caminho_pasta = "./leitores");
}
