% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baixar_pnadc_anual.R
\name{baixar_pnadc_anual}
\alias{baixar_pnadc_anual}
\title{Baixa e descompacta os microdados da PNADC anual}
\usage{
baixar_pnadc_anual(ano, caminho_pasta)
}
\arguments{
\item{ano}{Ano da PNADC Anual}

\item{caminho_pasta}{A pasta (o caminho para a pasta) na qual os arquivos serão guardados (se não existir, a pasta será criada)}
}
\value{
Os arquivos de microdados (txt) na pasta designada
}
\description{
Baixa e descompacta os microdados da PNADC anual (é preciso acesso a internet)
}
\examples{
baixar_pnadc_anual(ano = 2019, caminho_pasta = "./microdata");
}
