% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular_desocupados_anual.R
\name{calcular_desocupados_anual}
\alias{calcular_desocupados_anual}
\title{Calculate the unoccupied population}
\usage{
calcular_desocupados_anual(df_pnadc)
}
\arguments{
\item{df_pnadc}{The dataframe with the Annual PNADC data (see ler_pnadc_anual)}
}
\value{
A dataframe with 3 cols: year, total unoccupied and % of the workforce that are unoccupied
}
\description{
Calculate the unoccupied population
}
\examples{
calcular_desocupados_anual(pnadc_2019);
}
