% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular_nem_nem_anual.R
\name{calcular_nem_nem_anual}
\alias{calcular_nem_nem_anual}
\title{Calculate the "nem-nem" youth population}
\usage{
calcular_nem_nem_anual(df_pnad, idade_min = 15, idade_max = 29)
}
\arguments{
\item{idade_min}{The (inclusive) lowerbound age to define youth - default = 15}

\item{idade_max}{The (inclusive) upperbound age to define youth - deafult = 29}

\item{df_pnadc}{The dataframe with the Annual PNADC data (see ler_pnadc_anual)}
}
\value{
A dataframe with 3 cols: year, total youth "nem-nem" and % of the youth that are "nem-nem"
}
\description{
Calculate the "nem-nem" youth population: people between "idade_min" and "idade_max" that neither study nor work
}
\examples{
calcular_desocupados_anual(pnadc_2019);
}
\references{
Neri, 2021 - https://www.cps.fgv.br/cps/NemNem/
}
