% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular_renda_trabalho_media_anual.R
\name{calcular_renda_trabalho_media_anual}
\alias{calcular_renda_trabalho_media_anual}
\title{Calculate average work income}
\usage{
calcular_renda_trabalho_media_anual(df_pnadc, tipo)
}
\arguments{
\item{df_pnadc}{The dataframe with the Annual PNADC data (see ler_pnadc_anual)}

\item{tipo}{"habitual" or "efetivo" The kind of income: "habitual" is the customary monthly income; "efetivo" is the income actually received in that month}
}
\value{
A dataframe with 3 cols: year, real average monthly work income and nominal average monthly work income
}
\description{
Calculate average work income
}
\examples{
calcular_renda_trabalho_media_anual(pnadc_2019, tipo = "habitual");
calcular_renda_trabalho_media_anual(pnadc_2013, tipo = "efetivo");
}
