% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular_renda_trabalho_percentil_anual.R
\name{calcular_renda_trabalho_percentil_anual}
\alias{calcular_renda_trabalho_percentil_anual}
\title{Calculate percentile work income}
\usage{
calcular_renda_trabalho_percentil_anual(df_pnadc, percentil = 0.5, tipo)
}
\arguments{
\item{df_pnadc}{The dataframe with the Annual PNADC data (see ler_pnadc_anual)}

\item{percentil}{Numeric. The percentile (default = 0.5 : median). Can be more than one (see examples)}

\item{tipo}{"habitual" or "efetivo" The kind of income: "habitual" is the customary monthly income; "efetivo" is the income actually received in that month}
}
\value{
A dataframe with 3 cols: year, percentile and the respective monthly work real income
}
\description{
Calculate percentile work income
}
\examples{
calcular_renda_trabalho_percentil_anual(pnadc_2019);
calcular_renda_trabalho_percentil_anual(pnadc_2015, percentil = c(0.1, 0.4, 0.9, 0.95))
}
