% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{deflator}
\alias{deflator}
\title{Annual PNADC deflator}
\format{
A data frame with 972 rows and 6 variables:
\describe{
  \item{ID_deflator}{key value combining Year (4 digits), Quarter (1 digit) and UF (2 digits), to join with microdata}
  \item{CO1}{deflator for the same year, usual (habitual) values}
  \item{CO1e}{deflator for the same year, effective (efetivo) values}
  \item{CO2}{deflator for the last year, usual (habitual) values}
  \item{CO2e}{deflator for the last year, effective (efetivo) values}
  \item{CO3}{deflator for calculations related to poverty lina}
}
}
\usage{
deflator
}
\description{
Annual PNADC deflator
}
\seealso{
https://ftp.ibge.gov.br/Trabalho_e_Rendimento/Pesquisa_Nacional_por_Amostra_de_Domicilios_continua/Anual/Microdados/Visita/Documentacao_Geral/PNADcIBGE_Deflator_Anual_Visita.pdf
}
\keyword{datasets}
