% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{leitores_2013}
\alias{leitores_2013}
\title{Leitores da PNADC anual - 2013}
\format{
Um dataframe com 247 linhas e 4 variáveis
\describe{
  \item{inicio}{numeric, posição inicial da variável no arquivo txt de microdados}
  \item{fim}{numeric, posição final da variável no arquivo txt de microdados}
  \item{tamanho}{numeric, tamanho da variável no arquivo txt de microdados}
  \item{variavel}{character, nome da variável no arquivo txt de microdados}
}
}
\usage{
leitores_2013
}
\description{
Leitores da PNADC anual - 2013
}
\keyword{datasets}
