% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_pnadc_anual.R
\name{ler_pnadc_anual}
\alias{ler_pnadc_anual}
\title{Load annual PNADC microdata}
\usage{
ler_pnadc_anual(ano, path_pnadc)
}
\arguments{
\item{ano}{Year}

\item{path_pnadc}{The folder in which the microdata txt files are stores}
}
\value{
A dataframe with all the microdata information
}
\description{
Load annual PNADC microdata
}
\examples{
ler_pnadc_anual(2018, "./microdata_folder");
}
