% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montar_tabela_pobreza_anual.R
\name{montar_tabela_pobreza_anual}
\alias{montar_tabela_pobreza_anual}
\title{Add monetary poverty condition as a column}
\usage{
montar_tabela_pobreza_anual(df_pnadc, linha_pob)
}
\arguments{
\item{df_pnadc}{The dataframe with the Annual PNADC data (see ler_pnadc_anual)}

\item{linha_pob}{The monetary poverty threshold (household per capita monthly income)}
}
\value{
A dataframe without rows in which VD2002 is 15, 16, 17; and +2 cols: rdpc with household per capita monthly income and pobreza that equals 1 if person is poor (household per capita monthly income lower than the poverty threshold)
}
\description{
Add household per capita monthly income as a column
}
\examples{
montar_tabela_pobreza_anual(pnadc_2019, 400);
}
