% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montar_tabela_rdpc_anual.R
\name{montar_tabela_rdpc_anual}
\alias{montar_tabela_rdpc_anual}
\title{Add household per capita monthly income as a column}
\usage{
montar_tabela_rdpc_anual(df_pnadc)
}
\arguments{
\item{df_pnadc}{The dataframe with the Annual PNADC data (see ler_pnadc_anual)}
}
\value{
A dataframe without rows in which VD2002 is 15, 16, 17; and +1 col (rdpc) with household per capita monthly income
}
\description{
Add household per capita monthly income as a column
}
\examples{
montar_tabela_rdpc_anual(pnadc_2019);
}
